// src/app/interfaces/AdjuntoInfo.ts
export interface AdjuntoInfo {
    id?: string;
    codi: string;
    breq_codi: string; 
    nombre: string;
    desc: string;
    peso_adjunto: string; 
    extension: string;
    url: string;
    ind_estado: string;
    n_descargas: string; 
    fechaActividad: string;
    idUsuario: string; 
  }