// src/app/interfaces/Asignacion/AsignarTicket.ts
export interface AsignarTicket {
  id?: string;
  codi: string;
  nombre: string;
  breq_codi: string;
  klog_codi: string;
  prog_codi: string;
  cate_codi: string;
  fecha_inicio: string;
  fecha_fin: string;
  descripcion: string;
  fecha_actividad: string;
  ind_estado: string;
  idUsuario: string;
}